Bringing it together! Wsfy for the expr stmts, not assg exprs.
The worksheet is a little bare if it only instrumented assignment
expressions, since this didn't then include function calls, or just
"what's the value of this variable now". e.g. "x;" statements.

Another side effect is that assignments even in "for(;; i = i+1)" would
be instrumented, which isn't good.

Tidied up the constructor to TypeInference, and changed for
Instrumentor, to allow for the type inference.

Also. TypeInference is now work-in-progress transitioning so that the
CType of e.g. "5" is Primitive("5", "int"). This is partly because the
ST4 template uses the id ("5") to output the value.
May have other benefits.