Type Specifiers now handled more carefully.
This was a big pain with e.g.:
  typedef int T;
  { typedef float T; }
where the second typedef's T doesn't get parsed as a declarator, since
it's a typedefName (typeSpecifier).

The changes to the grammar are a bit untidy in 'ensuring' that
a declaration has at least a struct/union tag, enum constants, or
a declarator.
It could be tidied up, I think, but at least the above testcase works.