Move logic to 'CTypeFromDeclaration' from 'StringConstruction'.
'StringConstruction' is a bad name; but also, the class was doing the
work of both an AST Listener, as well as having this logic.

Another result of this was that classes like Insturmentor or
TypeInference, which happen to need Scopes also, 'hid' this dependency
by going through StringConstruction.

'StringConstruction' is still a bad name for the phase/pass of the AST,
though.