Instrumentor now adds typedefs from headers. Not super quick.
The program is 'slow' for now, because it's "doing a lot", and a fair
amount is redundant.

First, it parses the input program for a list of include headers.
(Cheap, but also not using a 'proper' C-PreProcessor. No parsing, only
goes through the Lexer).

Secondly, it parses these header files, extracting the typedefNames from
these headers. (e.g. the `size_t` in `typedef int size_t;`).
This is so the CParser can parse declarations like `size_t x;` properly.

Once the input program is then parsed:
Thirdly, it extracts the typedefs from these same header files,
and adds them to the GlobalScope (attached to the AST).
This is so we can generate strings for e.g. `size_t x;`.