Worksheet will kill a program after max # iterations.
This 'solves' the problem of infinite loops, at the expense of
not running programs which have a large number of iterations.
(Of course, in general it's hard to distinguish between "won't
terminate" and "long running program which terminates").

In this case, the default number of iterations is a low 1000.
I don't know what a sensible default value is. In this case, it's
because 10k took ~3-4 seconds to terminate on a trivial infinite loop
program.
It would be good if there were some way to specify max # iterations.