Add a proguardJar task to build.gradle.
This builds a jar which is shrunk/optimised/obfuscated using Proguard.
The build script makes use of shadow to achieve this, since it's an
easy way of dealing with duplicate class definitions in dependency jars.

There is the non-portable assumption that proguard is in
`/usr/share/proguard`. This should instead be a Gradle setting?